Move the cast to CLAMPOFFSET to get better format.

+++ tools/clang/unittests/HLSLExec/ExecutionTest.cpp
@@ -4057,7 +4057,7 @@ TEST_F(ExecutionTest, ATOWriteMSAATest) {
 }
 
 // Used to determine how an out of bounds offset should be converted
-#define CLAMPOFFSET(offset) ((offset<<28)>>28)
+#define CLAMPOFFSET(offset) (((unsigned)(offset)<<28)>>28)
 
 // Determine if the values in pPixels correspond to the expected locations encoded into a uint
 // based on the coordinates and offsets that were provided.
@@ -4066,24 +4066,9 @@ void VerifyProgOffsetResults(unsigned *pPixels, bool bCheckDeriv) {
   unsigned ix = 0;
   int coords[18] = {100, 150, 200, 250, 300, 350, 400, 450, 500, 550, 600, 650, 700, 750, 800, 850, 900, 950};
 
-  int offsets[18] = {CLAMPOFFSET((unsigned)-9),
-                     -8,
-                     -7,
-                     -6,
-                     -5,
-                     -4,
-                     -3,
-                     -2,
-                     -1,
-                     0,
-                     1,
-                     2,
-                     3,
-                     4,
-                     5,
-                     6,
-                     7,
-                     CLAMPOFFSET(8)};
+  int offsets[18] = {
+      CLAMPOFFSET(-9), -8, -7, -6, -5, -4, -3, -2, -1, 0, 1, 2, 3, 4, 5, 6, 7,
+      CLAMPOFFSET(8)};
 
   for (unsigned y = 0; y < _countof(coords); y++) {
     for (unsigned x = 0; x < _countof(coords); x++) {
